requirements.txt
torch==1.10.0
torchaudio==0.10.0
torchsummary
torchaudio-augmentations
torchvision
tensorboard
Pillow
numpy
tqdm
librosa
pandas
pytest
pyyaml
matplotlib
seaborn
ipython
scipy
scikit-learn
performer_pytorch
einops
